Separate events and event types.
